Redefine network loads
Each container knows how much traffic goes to the specific container. The node knows how much traffic goes through the node including the traffic going to it's containers. The difference between these 2 values is the sum of the traffic terminating at the node and the traffic being routed through the node.